python3-requirements.txt
PyYAML
paho-mqtt
gevent
gevent-websocket
